requirements.txt
Django==3.0.6
djangorestframework==3.11.0
requests==2.23.0